refactor: Extracted Movement behaviour from Ant
