+++ Engine/Entities/LivingEntity.cs
@@ -1,3 +1,5 @@
+using System;
+using System.Numerics;
 using AntEngine.Entities.States;
 using AntEngine.Entities.States.Living;
 using AntEngine.Utils.Maths;
@@ -9,7 +11,8 @@ namespace AntEngine.Entities
     /// </summary>
     public abstract class LivingEntity : StateEntity
     {
-        protected int _health;
+        private int _health;
+        private float _speed;
 
         public LivingEntity(World world) : this("Living Entity", new Transform(), world)
         {
@@ -39,6 +42,40 @@ namespace AntEngine.Entities
         /// </summary>
         public int MaxHealth { get; protected set; }
 
+        /// <summary>
+        /// Current speed of the ant.
+        /// </summary>
+        public float Speed
+        {
+            get => _speed;
+            protected set => _speed = value > MaxSpeed ? MaxSpeed : value;
+        }
+
+        /// <summary>
+        /// Maximum speed of the ant.
+        /// </summary>
+        public float MaxSpeed { get; protected set; }
+        
+        /// <summary>
+        /// Applies movement to ant's coordinates.
+        /// </summary>
+        /// <param name="dir"></param>
+        public void Move(Vector2 dir)
+        {
+            Vector2 globalDir = dir.Length() > 0
+                ? Vector2.Normalize(new Vector2(
+                    dir.X * MathF.Cos(Transform.Rotation) - dir.Y * MathF.Sin(Transform.Rotation),
+                    dir.X * MathF.Sin(Transform.Rotation) + dir.Y * MathF.Cos(Transform.Rotation)))
+                : Vector2.Zero;
+
+            Transform.Rotation = MathF.Atan2(globalDir.Y, Vector2.Dot(globalDir, Vector2.UnitX));
+
+            Vector2 lastPos = Transform.Position;
+            Collider.ParentTransform.Position += globalDir * Speed;
+            if (Collider.CheckCollision(World.Collider)) Collider.ParentTransform.Position = lastPos;
+        }
+
+        
         /// <summary>
         /// Kills the entity.
         /// </summary>
